Added ability to see user information including balance

+++ db/create.sql
@@ -4,7 +4,7 @@ CREATE TABLE Users (
     password VARCHAR(255) NOT NULL,
     firstname VARCHAR(255) NOT NULL,
     lastname VARCHAR(255) NOT NULL,
-    balance INT DEFAULT 0.0
+    balance DECIMAL(12,2) DEFAULT 0.0
 );
 
 CREATE TABLE Products (
